+++ internal/interface/graphql/factory/place.go
@@ -2,7 +2,7 @@ package factory
 
 import (
 	"poroto.app/poroto/planner/internal/domain/models"
-	graphql "poroto.app/poroto/planner/internal/infrastructure/graphql/model"
+	graphql "poroto.app/poroto/planner/internal/interface/graphql/model"
 )
 
 func PlaceFromDomainModel(place *models.Place) *graphql.Place {
