+++ internal/interface/graphql/factory/plan.go
@@ -3,7 +3,7 @@ package factory
 import (
 	"fmt"
 	"log"
-	graphql "poroto.app/poroto/planner/internal/infrastructure/graphql/model"
+	graphql "poroto.app/poroto/planner/internal/interface/graphql/model"
 
 	"poroto.app/poroto/planner/internal/domain/models"
 )
