+++ internal/interface/graphql/factory/price_range.go
@@ -2,7 +2,7 @@ package factory
 
 import (
 	"poroto.app/poroto/planner/internal/domain/models"
-	graphql "poroto.app/poroto/planner/internal/infrastructure/graphql/model"
+	graphql "poroto.app/poroto/planner/internal/interface/graphql/model"
 )
 
 func PriceRangeFromDomainModel(priceRange *models.PriceRange) *graphql.PriceRange {
