+++ internal/interface/graphql/resolver/plan_query.resolvers.go
@@ -12,8 +12,8 @@ import (
 	"poroto.app/poroto/planner/internal/domain/models"
 	"poroto.app/poroto/planner/internal/domain/services/plan"
 	"poroto.app/poroto/planner/internal/domain/services/user"
-	factory2 "poroto.app/poroto/planner/internal/infrastructure/graphql/factory"
-	"poroto.app/poroto/planner/internal/infrastructure/graphql/model"
+	"poroto.app/poroto/planner/internal/interface/graphql/factory"
+	"poroto.app/poroto/planner/internal/interface/graphql/model"
 )
 
 // Plan is the resolver for the plan field.
@@ -32,7 +32,7 @@ func (r *queryResolver) Plan(ctx context.Context, id string) (*model.Plan, error
 		return nil, nil
 	}
 
-	graphqlPlan, err := factory2.PlanFromDomainModel(*p, nil)
+	graphqlPlan, err := factory.PlanFromDomainModel(*p, nil)
 	if err != nil {
 		log.Println(err)
 		return nil, fmt.Errorf("internal server error")
@@ -55,7 +55,7 @@ func (r *queryResolver) Plans(ctx context.Context, pageKey *string) ([]*model.Pl
 		return nil, fmt.Errorf("could not fetch plans")
 	}
 
-	return factory2.PlansFromDomainModel(plans, nil), nil
+	return factory.PlansFromDomainModel(plans, nil), nil
 }
 
 // PlansByLocation is the resolver for the plansByLocation field.
@@ -81,7 +81,7 @@ func (r *queryResolver) PlansByLocation(ctx context.Context, input model.PlansBy
 	}
 
 	return &model.PlansByLocationOutput{
-		Plans:   factory2.PlansFromDomainModel(plans, nil),
+		Plans:   factory.PlansFromDomainModel(plans, nil),
 		PageKey: nextPageToken,
 	}, nil
 }
@@ -117,7 +117,7 @@ func (r *queryResolver) PlansByUser(ctx context.Context, input model.PlansByUser
 	}
 
 	return &model.PlansByUserOutput{
-		Plans:  factory2.PlansFromDomainModel(plans, nil),
-		Author: factory2.UserFromDomainModel(author),
+		Plans:  factory.PlansFromDomainModel(plans, nil),
+		Author: factory.UserFromDomainModel(author),
 	}, nil
 }
