+++ internal/interface/graphql/schema/image_type.graphqls
