+++ internal/interface/graphql/schema/place_type.graphqls
