+++ internal/interface/graphql/schema/plan_candidate_mutation.graphqls
