+++ internal/interface/graphql/schema/plan_candidate_query.graphqls
