+++ internal/interface/graphql/schema/plan_candidate_type.graphqls
