+++ internal/interface/graphql/schema/plan_query.graphqls
