+++ internal/interface/graphql/schema/plan_type.graphqls
