+++ internal/interface/graphql/schema/user_query.graphqls
