+++ internal/interface/graphql/schema/user_type.graphqls
